Codigo/QAE/qae_publisher/requirements.txt
functions-framework==3.*
requests
gspread
pydrive
pandas
numpy
google-cloud-secret-manager